use unified method to handle errors encountered during portforward
